requirements.txt
dill==0.3.8
dominate==2.9.1
gdown==5.1.0
ignite==1.1.0
lpips==0.1.4
matplotlib==3.5.0
mlflow==2.10.2
monai-weekly==1.2.dev2304
nibabel==4.0.2
numpy
omegaconf==2.3.0
pandas==2.2.0
parameterized==0.9.0
pillow
pynvml==11.5.0
scikit_learn==1.4.0
scipy==1.12.0
setuptools
scikit-image
statsmodels==0.14.1
tensorboardX==2.6.2.2
tensorflow==2.15.0.post1
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
x_transformers==1.27.16
